Refactor PersistentListSerializer tests
