+++ y-core/src/commonMain/kotlin/com.github.whyrising.y/collections/concretions/serialization/PersistentListSerializer.kt
@@ -1,0 +1,24 @@
+package com.github.whyrising.y.collections.concretions.serialization
+
+import com.github.whyrising.y.collections.concretions.list.PersistentList
+import com.github.whyrising.y.toPlist
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.builtins.ListSerializer
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
+
+// TODO: 4/4/22 Add documentation of O(n) complexity for this serializer.
+internal class PersistentListSerializer<E>(element: KSerializer<E>) :
+    KSerializer<PersistentList<E>> {
+
+    internal val listSerializer = ListSerializer(element)
+
+    override val descriptor: SerialDescriptor = listSerializer.descriptor
+
+    override fun deserialize(decoder: Decoder): PersistentList<E> =
+        listSerializer.deserialize(decoder).toPlist()
+
+    override fun serialize(encoder: Encoder, value: PersistentList<E>) =
+        listSerializer.serialize(encoder, value)
+}
